refactor: modularize PreviewPanel into composable components
- Decompose monolithic PreviewPanel into focused, reusable components
- Extract PlaybackControls, SelectionOverlay, HighlightMaskRenderer, VideoCanvas, PreviewHeaderControls, and SamplingViews
- Improve component organization and testability
- Remove original monolithic PreviewPanel.swift
